+++ Repostitory/Implementation/RentalItemRepostitory.cs
@@ -1,10 +1,12 @@
+using AutoMapper;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.EntityFrameworkCore;
 using WinterEquipmentRentalApi.Models;
 using WinterEquipmentRentalApi.Repostitory.Abstraction;
 
 namespace WinterEquipmentRentalApi.Repostitory.Implementation;
 
-public class RentalItemRepostitory(RentDbContext context) : IRentalItemRepository
+public class RentalItemRepostitory(RentDbContext context, IMapper mapper) : IRentalItemRepository
 {
     public async Task<string> Add(RentalItem entity)
     {
@@ -22,33 +24,37 @@ public class RentalItemRepostitory(RentDbContext context) : IRentalItemRepositor
 
     public async Task<RentalItem?> GetById(string id)
     {
-        return await context.RentalItems.FindAsync(id) 
-            ?? throw new Exception("Cannot find rental item");
+        return await context.RentalItems.FindAsync(id);
     }
 
-    public async Task<RentalItem> GetByName(string name)
+    public async Task<RentalItem?> GetByName(string name)
     {
         return await context.RentalItems.FirstAsync(x => x.Name == name);
     }
 
-    public async Task Remove(string id)
+    public async Task<bool> Remove(string id)
     {
-        var item = await context.RentalItems.FindAsync(id)
-            ?? throw new Exception("Cannot find rental item");
+        var item = await context.RentalItems.FindAsync(id);
+
+        if (item == null) return false;
 
         context.RentalItems.Remove(item);
-    }
-
-    public async Task Update(RentalItem entity)
-    {
-        var item = await context.RentalItems.FindAsync(entity.Id)
-            ?? throw new Exception("Cannot find rental item");
-
-        item.Id = entity.Id;
-        item.Name = entity.Name;
-        item.IsAvailable = entity.IsAvailable;
-        item.Price = entity.Price;
 
         await context.SaveChangesAsync();
+
+        return true;
+    }
+
+    public async Task<bool> Update(RentalItem entity)
+    {
+        var item = await context.RentalItems.FindAsync(entity.Id);
+
+        if (item == null) return false;
+
+        mapper.Map(entity, item);
+
+        await context.SaveChangesAsync();
+        
+        return true;
     }
 }
